+++ tests/sksl/shared/ResizeMatrix.metal
@@ -2,13 +2,13 @@
 #include <simd/simd.h>
 using namespace metal;
 struct Uniforms {
-    float4 colorGreen;
-    float4 colorRed;
+    half4 colorGreen;
+    half4 colorRed;
 };
 struct Inputs {
 };
 struct Outputs {
-    float4 sk_FragColor [[color(0)]];
+    half4 sk_FragColor [[color(0)]];
 };
 float2x2 float2x2_from_float3x3(float3x3 x0) {
     return float2x2(float2(x0[0].xy), float2(x0[1].xy));
