+++ tests/sksl/shared/Structs.metal
@@ -13,7 +13,7 @@ struct B {
 struct Inputs {
 };
 struct Outputs {
-    float4 sk_FragColor [[color(0)]];
+    half4 sk_FragColor [[color(0)]];
 };
 struct Globals {
     A a1;
@@ -26,6 +26,6 @@ fragment Outputs fragmentMain(Inputs _in [[stage_in]], bool _frontFacing [[front
     (void)_out;
     _globals.a1.x = 0;
     _globals.b1.x = 0.0;
-    _out.sk_FragColor.x = float(_globals.a1.x) + _globals.b1.x;
+    _out.sk_FragColor.x = half(_globals.a1.x) + half(_globals.b1.x);
     return _out;
 }
